fix relative imports after moving src
