+++ ui-react/src/types.tsx
@@ -1,7 +1,7 @@
 /// ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 // Types common with the electron code
 
-import { Credentials } from '../../src/backend/commonTypes';
+import { Credentials } from '../../src/src/backend/commonTypes';
 
 /// ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 export enum OutputVendorName {
